[FIX] Survive invalid context.
In some cases xmlrpc calls to the search method result in a value for
context that is True. This is because the search method does not have
the context as a last argument and xmlrpc does not handle keyword
parameters.

This commit replaces an invalid context with {}, to be able to use
xmlrpc based tools like Pentaho.

I was nt able to find out how to fix this in a better way.